requirements.txt
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
tensorboard==2.15.1
pyyaml==6.0.1
tqdm==4.66.1
jupyter==1.0.0
pandas==2.1.4
pyarrow==14.0.2
